chore(lint): add prettier + solhint + pre-commit hooks

+++ src/VowJoin.sol
@@ -56,12 +56,13 @@ contract VowJoin {
 }
 
 interface DaiLike {
-    function approve(address, uint) external;
+    function approve(address, uint256) external;
+
     function balanceOf(address) external view returns (uint256);
 }
 
 interface DaiJoinLike {
     function dai() external view returns (address);
+
     function join(address, uint256) external;
 }
-
